split search panel configs
